Reworked the implementation of DFAs
Instead of lambda functions I'm using transition tables as a dict object. I also added the ability to set a fallback key '_default'. When there isn't an explicit transition from one state to another with a certain character then it will go to the fallback state. This is useful for DFAs with a large Σ