Fixed get_accepted() when it was re-wrote
Forgot to return a String with the empty 'epsilon' character when the initial state is an accepting state. Before I just returned a String constructed with an empty list which when compared in the subset function would cause acceptable empty strings to evaluate to False. Now acceptable empty strings are returned as String([Char()])