Completed Task #32
(Backtracking) Write a function that given an NFA and a string, determines if the string is accepted.

I'm not to sure if this is how Jay wanted it implemented. He might think it's a 'lame way'. But the previous version of this function has an issue is we the initial state is added to the epslion closure stack even if it doesnt have an epsilon transition.

Might come back to this